Module kiss3d::builtin

Built-in geometries, shaders and effects.

Structs

NormalsMaterial
A material that draws normals of an object.

ObjectMaterial
The default material used to draw objects.

PlanarObjectMaterial
The default material used to draw objects.

UvsMaterial
A material that draws normals of an object.

Statics

NORMAL_FRAGMENT_SRC
A fragment shader for coloring each point of an object depending on its normal.

NORMAL_VERTEX_SRC
A vertex shader for coloring each point of an object depending on its normal.

OBJECT_FRAGMENT_SRC
Fragment shader of the default object material.

OBJECT_VERTEX_SRC
Vertex shader of the default object material.

UVS_FRAGMENT_SRC
A fragment shader for coloring each point of an object depending on its texture coordinates.

UVS_VERTEX_SRC
A vertex shader for coloring each point of an object depending on its texture coordinates.