Struct kiss3d::camera::ArcBall

pub struct ArcBall { /* fields omitted */ }

Arc-ball camera mode.

An arc-ball camera is a camera rotating around a fixed point (the focus point) and always looking at it. The following inputs are handled:

Left button press + drag - rotates the camera around the focus point

Right button press + drag - translates the focus point on the plane orthogonal to the view direction

Scroll in/out - zoom in/out

Enter key - set the focus point to the origin

Implementations

impl ArcBall

pub fn new(eye: Point3<f32>, at: Point3<f32>) -> ArcBall

Create a new arc-ball camera.

pub fn new_with_frustrum(
fov: f32,
znear: f32,
zfar: f32,
eye: Point3<f32>,
at: Point3<f32>
) -> ArcBall

Creates a new arc ball camera with default sensitivity values.

pub fn at(&self) -> Point3<f32>

The point the arc-ball is looking at.

pub fn set_at(&mut self, at: Point3<f32>)

Get a mutable reference to the point the camera is looking at.

pub fn yaw(&self) -> f32

The arc-ball camera yaw.

pub fn set_yaw(&mut self, yaw: f32)

Sets the camera yaw. Change this to modify the rotation along the up axis.

pub fn pitch(&self) -> f32

The arc-ball camera pitch.

pub fn set_pitch(&mut self, pitch: f32)

Sets the camera pitch.

pub fn min_pitch(&self) -> f32

The minimum pitch of the camera.

pub fn set_min_pitch(&mut self, min_pitch: f32)

Set the minimum pitch of the camera.

pub fn max_pitch(&self) -> f32

The maximum pitch of the camera.

pub fn set_max_pitch(&mut self, max_pitch: f32)

Set the maximum pitch of the camera.

pub fn dist(&self) -> f32

The distance from the camera position to its view point.

pub fn set_dist(&mut self, dist: f32)

Move the camera such that it is at a given distance from the view point.

pub fn min_dist(&self) -> f32

The minimum distance from the camera position to its view point.

pub fn set_min_dist(&mut self, min_dist: f32)

Set the minimum distance from the camera position to its view point.

pub fn max_dist(&self) -> f32

The maximum distance from the camera position to its view point.

pub fn set_max_dist(&mut self, max_dist: f32)

Set the maximum distance from the camera position to its view point.

pub fn look_at(&mut self, eye: Point3<f32>, at: Point3<f32>)

Move and orient the camera such that it looks at a specific point.

pub fn rotate_button(&self) -> Option<MouseButton>

The button used to rotate the ArcBall camera.

pub fn rebind_rotate_button(&mut self, new_button: Option<MouseButton>)

Set the button used to rotate the ArcBall camera. Use None to disable rotation.

pub fn rotate_modifiers(&self) -> Option<Modifiers>

Modifiers that must be pressed for the camera rotation to occur.

pub fn set_rotate_modifiers(&mut self, modifiers: Option<Modifiers>)

Sets the modifiers that must be pressed for the camera rotation to occur.

If this is set to None, then pressing any modifier will not prevent rotation from occurring. If this is different from None then rotation will occur only if the exact specified set of modifiers is pressed. In particular, if this is set to Some(Modifiers::empty()) then, rotation will occur only of no modifier is pressed.

pub fn drag_modifiers(&self) -> Option<Modifiers>

Modifiers that must be pressed for the camera drag to occur.

pub fn set_drag_modifiers(&mut self, modifiers: Option<Modifiers>)

Sets the modifiers that must be pressed for the camera drag to occur.

If this is set to None, then pressing any modifier will not prevent dragging from occurring. If this is different from None then drag will occur only if the exact specified set of modifiers is pressed. In particular, if this is set to Some(Modifiers::empty()) then, drag will occur only of no modifier is pressed.

pub fn drag_button(&self) -> Option<MouseButton>

The button used to drag the ArcBall camera.

pub fn rebind_drag_button(&mut self, new_button: Option<MouseButton>)

Set the button used to drag the ArcBall camera. Use None to disable dragging.

pub fn reset_key(&self) -> Option<Key>

The key used to reset the ArcBall camera.

pub fn rebind_reset_key(&mut self, new_key: Option<Key>)

Set the key used to reset the ArcBall camera. Use None to disable reset.

pub fn set_up_axis(&mut self, up_axis: Vector3<f32>)

Sets the up vector of this camera. Prefer using set_up_axis_dir if your up vector is already normalized.

pub fn set_up_axis_dir(&mut self, up_axis: Unit<Vector3<f32>>)

Sets the up-axis direction of this camera.