Struct kiss3d::camera::FirstPerson

pub struct FirstPerson { /* fields omitted */ }

First-person camera mode.

Left button press + drag - look around

Right button press + drag - translates the camera position on the plane orthogonal to the view direction

Scroll in/out - zoom in/out

Methods

impl FirstPerson

fn new(eye: Point3<f32>, at: Point3<f32>) -> FirstPerson

Creates a first person camera with default sensitivity values.

fn new_with_frustrum(fov: f32,
znear: f32,
zfar: f32,
eye: Point3<f32>,
at: Point3<f32>)
-> FirstPerson

Creates a new first person camera with default sensitivity values.

fn set_move_step(&mut self, step: f32)

Sets the translational increment per arrow press.

The default value is 0.5.

fn set_pitch_step(&mut self, step: f32)

Sets the pitch increment per mouse movement.

The default value is 0.005.

fn set_yaw_step(&mut self, step: f32)

Sets the yaw increment per mouse movement.

The default value is 0.005.

fn move_step(&self) -> f32

Gets the translational increment per arrow press.

fn pitch_step(&self) -> f32

Gets the pitch increment per mouse movement.

fn yaw_step(&self) -> f32

Gets the yaw increment per mouse movement.

fn look_at(&mut self, eye: Point3<f32>, at: Point3<f32>)

Changes the orientation and position of the camera to look at the specified point.

fn at(&self) -> Point3<f32>

The point the camera is looking at.

fn rotate_button(&self) -> Option<MouseButton>

The button used to rotate the FirstPerson camera.

fn rebind_rotate_button(&mut self, new_button: Option<MouseButton>)

Set the button used to rotate the FirstPerson camera. Use None to disable rotation.

fn drag_button(&self) -> Option<MouseButton>

The button used to drag the FirstPerson camera.

fn rebind_drag_button(&mut self, new_button: Option<MouseButton>)

Set the button used to drag the FirstPerson camera. Use None to disable dragging.

fn up_key(&self) -> Option<Key>

The movement button for up.

fn down_key(&self) -> Option<Key>

The movement button for down.

fn left_key(&self) -> Option<Key>

The movement button for left.

fn right_key(&self) -> Option<Key>

The movement button for right.

fn rebind_up_key(&mut self, new_key: Option<Key>)

Set the movement button for up. Use None to disable movement in this direction.

fn rebind_down_key(&mut self, new_key: Option<Key>)

Set the movement button for down. Use None to disable movement in this direction.

fn rebind_left_key(&mut self, new_key: Option<Key>)

Set the movement button for left. Use None to disable movement in this direction.

fn rebind_right_key(&mut self, new_key: Option<Key>)

Set the movement button for right. Use None to disable movement in this direction.

fn unbind_movement_keys(&mut self)

Disable the movement buttons for up, down, left and right.

fn eye_dir(&self) -> Vector3<f32>

The direction this camera is looking at.

fn move_dir(&self,
up: bool,
down: bool,
right: bool,
left: bool)
-> Vector3<f32>

The direction this camera is being moved by the keyboard keys for a given set of key states.

fn translate_mut(&mut self, t: &Translation3<f32>)

Translates in-place this camera by t.

fn translate(&self, t: &Translation3<f32>) -> FirstPerson

Translates this camera by t.