Struct kiss3d​::line_renderer​::LineRenderer​

pub struct LineRenderer {
    // some fields omitted
}

Structure which manages the display of short-living lines.

Methods

impl LineRenderer

fn new() -> LineRenderer

Creates a new lines manager.

fn needs_rendering(&self) -> bool

Indicates whether some lines have to be drawn.

fn draw_line(&mut self, a: Vec3<GLfloat>, b: Vec3<GLfloat>, color: Vec3<GLfloat>)

Adds a line to be drawn during the next frame. Lines are not persistent between frames. This method must be called for each line to draw, and at each update loop iteration.

fn render(&mut self, pass: uint, camera: &mut Camera)

Actually draws the lines.