Module kiss3d​::point_renderer​

A batched point renderer.

Structs

PointRenderer
Structure which manages the display of short-living points.

Statics

pub static POINTS_FRAGMENT_SRC: &'static str = ANOTHER_VERY_LONG_STRING
Fragment shader used by the material to display line.

pub static POINTS_VERTEX_SRC: &'static str = A_VERY_LONG_STRING
Vertex shader used by the material to display line.