Struct kiss3d::post_processing::Waves

pub struct Waves { /* fields omitted */ }

An useless post-processing effect mainly to test that everything works correctly.

It deforms the displayed scene with a wave effect.

Methods

impl Waves

fn new() -> Waves

Creates a new Waves post processing effect.