Module kiss3d::renderer::line_renderer

A batched line renderer.

Structs

LineRenderer
Structure which manages the display of short-living lines.

Statics

LINES_FRAGMENT_SRC
Fragment shader used by the material to display line.

LINES_VERTEX_SRC
Vertex shader used by the material to display line.