Module kiss3d::resource

GPU resource managers

Re-exports

pub use crate::resource::material::Material;

pub use crate::resource::material::PlanarMaterial;

Modules

material
Trait implemented by materials.

Structs

Effect
Structure encapsulating a program.

FramebufferManager
A framebuffer manager. It is a simple to to switch between an off-screen framebuffer and the default (window) framebuffer.

GPUVec
A vector of elements that can be loaded to the GPU, on the RAM, or both.

MaterialManager
The material manager.

Mesh
Aggregation of vertices, indices, normals and texture coordinates.

MeshManager
The mesh manager.

OffscreenBuffers
OpenGL identifiers to an off-screen buffer.

PlanarMaterialManager
The material manager.

PlanarMesh
Aggregation of vertices, indices, normals and texture coordinates.

PlanarMeshManager
The mesh manager.

ShaderAttribute
Structure encapsulating an attribute.

ShaderUniform
Structure encapsulating an uniform variable.

Texture

TextureManager
The texture manager.

Enums

AllocationType
Allocation type of gpu buffers.

BufferType
Type of gpu buffer.

PrimitiveArray
An array of primitive types.

RenderTarget
The target to every rendering call.

TextureWrapping
Wrapping parameters for a texture.

Traits

GLPrimitive
Trait implemented by structures that can be uploaded to a uniform or contained by a gpu array.