Struct kiss3d​::resource​::FramebufferManager​

pub struct FramebufferManager {
    // some fields omitted
}

A framebuffer manager. It is a simple to to switch between an off-screen framebuffer and the default (window) framebuffer.

Methods

impl FramebufferManager

fn new() -> FramebufferManager

Creates a new framebuffer manager.

fn new_render_target(width: uint, height: uint) -> RenderTarget

Creates a new render target. A render target is the combination of a color buffer and a depth buffer.

fn screen() -> RenderTarget

Returns the render target associated with the screen.

fn select(&mut self, target: &RenderTarget)

Selects a specific render target