Struct kiss3d::resource::Mesh

pub struct Mesh {
    // some fields omitted
}

Aggregation of vertices, indices, normals and texture coordinates.

It also contains the GPU location of those buffers.

Methods

impl Mesh

fn new(coords: Vec<Pnt3<GLfloat>>, faces: Vec<Pnt3<GLuint>>, normals: Option<Vec<Vec3<GLfloat>>>, uvs: Option<Vec<Pnt2<GLfloat>>>, dynamic_draw: bool) -> Mesh

Creates a new mesh.

If the normals and uvs are not given, they are automatically computed.

fn from_trimesh(mesh: TriMesh3<GLfloat>, dynamic_draw: bool) -> Mesh

Creates a new mesh from a mesh descr.

In the normals and uvs are not given, they are automatically computed.

fn to_trimesh(&self) -> Option<TriMesh3<GLfloat>>

Creates a triangle mesh from this mesh.

fn new_with_gpu_vectors(coords: Arc<RwLock<GPUVector<Pnt3<GLfloat>>>>, faces: Arc<RwLock<GPUVector<Pnt3<GLuint>>>>, normals: Arc<RwLock<GPUVector<Vec3<GLfloat>>>>, uvs: Arc<RwLock<GPUVector<Pnt2<GLfloat>>>>) -> Mesh

Creates a new mesh. Arguments set to None are automatically computed.

fn bind_coords(&mut self, coords: &mut ShaderAttribute<Pnt3<GLfloat>>)

Binds this mesh vertex coordinates buffer to a vertex attribute.

fn bind_normals(&mut self, normals: &mut ShaderAttribute<Vec3<GLfloat>>)

Binds this mesh vertex normals buffer to a vertex attribute.

fn bind_uvs(&mut self, uvs: &mut ShaderAttribute<Pnt2<GLfloat>>)

Binds this mesh vertex uvs buffer to a vertex attribute.

fn bind_faces(&mut self)

Binds this mesh vertex uvs buffer to a vertex attribute.

fn bind(&mut self, coords: &mut ShaderAttribute<Pnt3<GLfloat>>, normals: &mut ShaderAttribute<Vec3<GLfloat>>, uvs: &mut ShaderAttribute<Pnt2<GLfloat>>)

Binds this mesh buffers to vertex attributes.

fn unbind(&self)

Unbind this mesh buffers to vertex attributes.

fn num_pts(&self) -> usize

Number of points needed to draw this mesh.

fn recompute_normals(&mut self)

Recompute this mesh normals.

fn faces<'a>(&'a self) -> &'a Arc<RwLock<GPUVector<Pnt3<GLuint>>>>

This mesh faces.

fn normals<'a>(&'a self) -> &'a Arc<RwLock<GPUVector<Vec3<GLfloat>>>>

This mesh normals.

fn coords<'a>(&'a self) -> &'a Arc<RwLock<GPUVector<Pnt3<GLfloat>>>>

This mesh vertex coordinates.

fn uvs<'a>(&'a self) -> &'a Arc<RwLock<GPUVector<Pnt2<GLfloat>>>>

This mesh texture coordinates.

fn compute_normals_array(coordinates: &[Pnt3<GLfloat>], faces: &[Pnt3<GLuint>]) -> Vec<Vec3<GLfloat>>

Computes normals from a set of faces.

fn compute_normals(coordinates: &[Pnt3<GLfloat>], faces: &[Pnt3<GLuint>], normals: &mut Vec<Vec3<GLfloat>>)

Computes normals from a set of faces.