Struct kiss3d​::text​::Font​

pub struct Font {
    // some fields omitted
}

A ttf font.

Methods

impl Font

fn new(path: &Path, size: i32) -> Rc<Font>

Loads a new ttf font.

fn texture_atlas(&self) -> GLuint

The opengl id to the texture atlas of this font.

fn atlas_dimensions(&self) -> Vec2<uint>

The dimensions of the texture atlas of this font.

fn glyphs<'a>(&'a self) -> &'a [Option<Glyph>]

The glyphs of the this font.

fn height(&self) -> i32

The height of this font.