Module kiss3d​::window​

The window, and things to handle the rendering loop and events.

Structs

Event
An event.

EventManager
A stand-alone object that provides an iterator though glfw events.

Events
An iterator through events.

Window
Structure representing a window and a 3D scene.