Struct kiss3d::window::Event

pub struct Event<'a> {
    pub timestamp: f64,
    pub value: WindowEvent,
    pub inhibited: bool,
    // some fields omitted
}

An event.

Fields
timestamp: f64

The event timestamp.

value: WindowEvent

The exact glfw event value. This can be modified to fool the other event handlers.

inhibited: bool

Set this to true to prevent the window or the camera from handling the event.