Struct kiss3d​::window​::EventManager​

pub struct EventManager {
    // some fields omitted
}

A stand-alone object that provides an iterator though glfw events.

It is not lifetime-bound to the main window.

Methods

impl EventManager

fn new(events: Rc<Receiver<(f64, WindowEvent)>>, inhibitor: Rc<RefCell<Vec<WindowEvent>>>) -> EventManager

Creates a new event manager.

fn iter<'a>(&'a mut self) -> Events<'a>

Gets an iterator to the glfw events already collected.