Struct kiss3d::camera::ArcBall

pub struct ArcBall {
    // some fields omitted
}

Arc-ball camera mode.

An arc-ball camera is a camera rotating around a fixed point (the focus point) and always looking at it. The following inputs are handled:

Left button press + drag - rotates the camera around the focus point

Right button press + drag - translates the focus point on the plane orthogonal to the view direction

Scroll in/out - zoom in/out

Enter key - set the focus point to the origin

Methods

impl ArcBall

fn new(eye: Pnt3<f32>, at: Pnt3<f32>) -> ArcBall

Create a new arc-ball camera.

fn new_with_frustrum(fov: f32, znear: f32, zfar: f32, eye: Pnt3<f32>, at: Pnt3<f32>) -> ArcBall

Creates a new arc ball camera with default sensitivity values.

fn at(&self) -> Pnt3<f32>

The point the arc-ball is looking at.

fn set_at(&mut self, at: Pnt3<f32>)

Get a mutable reference to the point the camera is looking at.

fn yaw(&self) -> f32

The arc-ball camera yaw.

fn set_yaw(&mut self, yaw: f32)

Sets the camera yaw. Change this to modify the rotation along the up axis.

fn pitch(&self) -> f32

The arc-ball camera pitch.

fn set_pitch(&mut self, pitch: f32)

Sets the camera pitch.

fn dist(&self) -> f32

The distance from the camera position to its view point.

fn set_dist(&mut self, dist: f32)

Move the camera such that it is at a given distance from the view point.

fn look_at_z(&mut self, eye: Pnt3<f32>, at: Pnt3<f32>)

Move and orient the camera such that it looks at a specific point.