Struct kiss3d::camera::ArcBall

pub struct ArcBall { /* fields omitted */ }

Arc-ball camera mode.

An arc-ball camera is a camera rotating around a fixed point (the focus point) and always looking at it. The following inputs are handled:

Left button press + drag - rotates the camera around the focus point

Right button press + drag - translates the focus point on the plane orthogonal to the view direction

Scroll in/out - zoom in/out

Enter key - set the focus point to the origin

Methods

impl ArcBall

fn new(eye: Point3<f32>, at: Point3<f32>) -> ArcBall

Create a new arc-ball camera.

fn new_with_frustrum(fov: f32,
znear: f32,
zfar: f32,
eye: Point3<f32>,
at: Point3<f32>)
-> ArcBall

Creates a new arc ball camera with default sensitivity values.

fn at(&self) -> Point3<f32>

The point the arc-ball is looking at.

fn set_at(&mut self, at: Point3<f32>)

Get a mutable reference to the point the camera is looking at.

fn yaw(&self) -> f32

The arc-ball camera yaw.

fn set_yaw(&mut self, yaw: f32)

Sets the camera yaw. Change this to modify the rotation along the up axis.

fn pitch(&self) -> f32

The arc-ball camera pitch.

fn set_pitch(&mut self, pitch: f32)

Sets the camera pitch.

fn dist(&self) -> f32

The distance from the camera position to its view point.

fn set_dist(&mut self, dist: f32)

Move the camera such that it is at a given distance from the view point.

fn look_at(&mut self, eye: Point3<f32>, at: Point3<f32>)

Move and orient the camera such that it looks at a specific point.

fn rotate_button(&self) -> Option<MouseButton>

The button used to rotate the ArcBall camera.

fn rebind_rotate_button(&mut self, new_button: Option<MouseButton>)

Set the button used to rotate the ArcBall camera. Use None to disable rotation.

fn drag_button(&self) -> Option<MouseButton>

The button used to drag the ArcBall camera.

fn rebind_drag_button(&mut self, new_button: Option<MouseButton>)

Set the button used to drag the ArcBall camera. Use None to disable dragging.

fn reset_key(&self) -> Option<Key>

The key used to reset the ArcBall camera.

fn rebind_reset_key(&mut self, new_key: Option<Key>)

Set the key used to reset the ArcBall camera. Use None to disable reset.