Struct kiss3d::camera::FirstPerson

pub struct FirstPerson {
    // some fields omitted
}

First-person camera mode.

Left button press + drag - look around

Right button press + drag - translates the camera position on the plane orthogonal to the view direction

Scroll in/out - zoom in/out

Methods

impl FirstPerson

fn new(eye: Pnt3<f32>, at: Pnt3<f32>) -> FirstPerson

Creates a first person camera with default sensitivity values.

fn new_with_frustrum(fov: f32, znear: f32, zfar: f32, eye: Pnt3<f32>, at: Pnt3<f32>) -> FirstPerson

Creates a new first person camera with default sensitivity values.

fn set_move_step(&mut self, step: f32)

Sets the translational increment per arrow press.

The default value is 0.5.

fn set_pitch_step(&mut self, step: f32)

Sets the pitch increment per mouse movement.

The default value is 0.005.

fn set_yaw_step(&mut self, step: f32)

Sets the yaw increment per mouse movement.

The default value is 0.005.

fn move_step(&self) -> f32

Gets the translational increment per arrow press.

fn pitch_step(&self) -> f32

Gets the pitch increment per mouse movement.

fn yaw_step(&self) -> f32

Gets the yaw increment per mouse movement.

fn look_at_z(&mut self, eye: Pnt3<f32>, at: Pnt3<f32>)

Changes the orientation and position of the camera to look at the specified point.

fn at(&self) -> Pnt3<f32>

The point the camera is looking at.

fn eye_dir(&self) -> Vec3<f32>

The direction this camera is looking at.

fn move_dir(&self, up: bool, down: bool, right: bool, left: bool) -> Vec3<f32>

The direction this camera is being moved by the keyboard keys for a given set of key states.