Struct kiss3d::camera::FirstPersonStereo

pub struct FirstPersonStereo { /* fields omitted */ }

First-person camera mode.

Left button press + drag - look around

Right button press + drag - translates the camera position on the plane orthogonal to the view direction

Scroll in/out - zoom in/out

Methods

impl FirstPersonStereo

fn new(eye: Point3<f32>, at: Point3<f32>, ipd: f32) -> FirstPersonStereo

Creates a first person camera with default sensitivity values.

fn new_with_frustrum(fov: f32,
znear: f32,
zfar: f32,
eye: Point3<f32>,
at: Point3<f32>,
ipd: f32)
-> FirstPersonStereo

Creates a new first person camera with default sensitivity values.

fn look_at(&mut self, eye: Point3<f32>, at: Point3<f32>)

Changes the orientation and position of the camera to look at the specified point.

fn at(&self) -> Point3<f32>

The point the camera is looking at.

fn ipd(&self) -> f32

return Inter Pupilary Distance

fn set_ipd(&mut self, ipd: f32)

change Inter Pupilary Distance