Struct kiss3d​::point_renderer​::PointRenderer​

pub struct PointRenderer {
    // some fields omitted
}

Structure which manages the display of short-living points.

Methods

impl PointRenderer

fn new() -> PointRenderer

Creates a new points manager.

fn needs_rendering(&self) -> bool

Indicates whether some points have to be drawn.

fn draw_point(&mut self, pt: Vec3<GLfloat>, color: Vec3<GLfloat>)

Adds a line to be drawn during the next frame. Lines are not persistent between frames. This method must be called for each line to draw, and at each update loop iteration.

fn render(&mut self, pass: uint, camera: &mut Camera)

Actually draws the points.