Module kiss3d​::post_processing​

Post-processing effects.

Reexports

pub use post_processing::post_processing_effect::PostProcessingEffect;

Modules

post_processing_effect
Trait implemented by every post-processing effect.

Structs

Grayscales
Post processing effect which turns everything in gray scales.

OculusStereo
An post-processing effect to support the oculus rift.

SobelEdgeHighlight
Post processing effect which turns everything in grayscales.

Waves
An useless post-processing effect mainly to test that everything works correctly.