Trait kiss3d​::post_processing​::post_processing_effect​::PostProcessingEffect​

pub trait PostProcessingEffect {
    fn update(&mut self, dt: f32, w: f32, h: f32, znear: f32, zfar: f32);
    fn draw(&mut self, target: &RenderTarget);
}

Trait of post processing effect.

One post-processing effect can be used at a time. It is executed once the scene has been rendered on a texture.

Required Methods

fn update(&mut self, dt: f32, w: f32, h: f32, znear: f32, zfar: f32)

Updates the post processing effect.

fn draw(&mut self, target: &RenderTarget)

Render the effect.

Arguments:

shader_manager - manager to switch between the different shaders.

fbo_texture - id to the texture containing the last scene drawn.

fbo_depth - the depth buffer as a texture.