Module kiss3d​::resource​

GPU resource managers

Reexports

pub use resource::material::Material;

pub use resource::gpu_vector::{StaticDraw, DynamicDraw, StreamDraw, ArrayBuffer, ElementArrayBuffer};

Modules

material
Trait implemented by materials.

Structs

FramebufferManager
A framebuffer manager. It is a simple to to switch between an off-screen framebuffer and the default (window) framebuffer.

GPUVector
A vector of elements that can be loaded to the GPU, on the RAM, or both.

MaterialManager
The material manager.

Mesh
Aggregation of vertices, indices, normals and texture coordinates.

MeshManager
The mesh manager.

OffscreenBuffers
OpenGL identifiers to an off-screen buffer.

Shader
Structure encapsulating a shader program.

ShaderAttribute
Structure encapsulating an attribute.

ShaderUniform
Structure encapsulating an uniform variable.

Texture
A gpu texture. It contains the texture id provided by opengl and is automatically released.

TextureManager
The texture manager.

Enums

AllocationType
Allocation type of gpu buffers.

BufferType
Type of gpu buffer.

RenderTarget
The target to every rendering call.

Traits

GLPrimitive
Trait implemented by structures that can be uploaded to a uniform or contained by a gpu array.