Trait kiss3d​::resource​::GLPrimitive​

pub trait GLPrimitive: Copy {
    fn gl_type(_type: Option<Self>) -> GLuint;
    fn size(_type: Option<Self>) -> GLuint;
    fn upload(&self, location: GLuint);
}

Trait implemented by structures that can be uploaded to a uniform or contained by a gpu array.

Required Methods

fn gl_type(_type: Option<Self>) -> GLuint

The opengl primitive type of this structure content.

fn size(_type: Option<Self>) -> GLuint

The number of elements of type self.gl_type() this structure stores.

fn upload(&self, location: GLuint)

Uploads the element to a gpu location.