Struct kiss3d::scene::ObjectData

pub struct ObjectData {
    // some fields omitted
}

Set of data identifying a scene node.

Methods

impl ObjectData

fn texture<'a>(&'a self) -> &'a Rc<Texture>

The texture of this object.

fn color<'a>(&'a self) -> &'a Pnt3<f32>

The color of this object.

fn lines_width(&self) -> f32

The width of the lines draw for this object.

fn points_size(&self) -> f32

The size of the points draw for this object.

fn surface_rendering_active(&self) -> bool

Whether this object has its surface rendered or not.

fn backface_culling_enabled(&self) -> bool

Whether this object uses backface culling or not.

fn user_data<'a>(&'a self) -> &'a Any

An user-defined data.

Use dynamic typing capabilities of the Any type to recover the actual data.