Struct kiss3d::text::Font

pub struct Font { /* fields omitted */ }

A ttf font.

Methods

impl Font

fn from_memory(font: &[u8], size: i32) -> Rc<Font>

Loads a new ttf font from the memory.

fn new(path: &Path, size: i32) -> Rc<Font>

Loads a new ttf font from a file.

fn do_new(path: Option<&Path>, memory: &[u8], size: i32) -> Rc<Font>

Loads a new ttf font from a file.

fn texture_atlas(&self) -> GLuint

The opengl id to the texture atlas of this font.

fn atlas_dimensions(&self) -> Vector2<usize>

The dimensions of the texture atlas of this font.

fn glyphs(&self) -> &[Option<Glyph>]

The glyphs of the this font.

fn height(&self) -> i32

The height of this font.